smaller changes while writting thesis

+++ plugins/de.cau.cs.kieler.comparison.compilers/src/de/cau/cs/kieler/comparison/compilers/kico/AbstractKiCo.java
@@ -93,7 +93,7 @@ public abstract class AbstractKiCo implements ICompiler {
      * {@inheritDoc}
      */
     @Override
-    public int getCompilationOffset() {
+    public long getCompilationOffset() {
         return 0;
     }
